Update to standard all .h files  correct condition in player_mov. pos_player()

+++ src/game/movement/mouse.c
@@ -27,7 +27,7 @@ void	use_mouse(t_game *game)
 	}
 	else if (delta < WIN_WIDTH * 0.5 - 1)
 	{
-		game->player->sensi = (float)-delta * 0.001;
+		game->player->sensi = -delta * 0.001;
 		rotate_right(game->player);
 	}
 	mlx_mouse_move(game->window->win_ptr, WIN_WIDTH * 0.5, WIN_HEIGHT * 0.5);
